backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for the Teknofest2025-TruthMark project"
authors = [{ name = "Zaman Kazimov", email = "kazimovzaman2@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.10.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "ruff>=0.12.5",
]
